// src/agents/swagger/enum/node-names.enum.ts
export enum NodeNames {
  SCAN_PROJECT = "scanProject",
  EXTRACT_ROUTER_CONTEXTS = "extractRouterContexts",
  ANALYZE_ROUTES = "analyzeRoutes",
  ANALYZE_CONTROLLERS = "analyzeControllers",
  MERGE_ROUTES_WITH_CONTROLLER = "mergeRoutesWithController",
  GENERATE_SWAGGER = "generateSwagger",
  UPDATE_ROUTES_WITH_DOCUMENTATION = "updateRoutesWithDocumentation",
  VALIDATE_OUTPUT = "validateOutput",
}